chore: Lost and found pagination API don't require auth token
