+++ data/src/main/java/in/koreatech/koin/data/api/auth/ArticleAuthApi.kt
@@ -4,7 +4,6 @@ import `in`.koreatech.koin.data.request.article.ArticleKeywordRequest
 import `in`.koreatech.koin.data.request.article.ArticleLostAndFoundReportRequest
 import `in`.koreatech.koin.data.request.article.ArticleLostAndFoundRequest
 import `in`.koreatech.koin.data.response.article.ArticleKeywordWrapperResponse
-import `in`.koreatech.koin.data.response.article.ArticleLostAndFoundPaginationResponse
 import `in`.koreatech.koin.data.response.article.ArticleLostAndFoundResponse
 import `in`.koreatech.koin.data.response.article.KeywordsResponse
 import retrofit2.Response
@@ -13,7 +12,6 @@ import retrofit2.http.DELETE
 import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.Path
-import retrofit2.http.Query
 
 interface ArticleAuthApi {
     @GET("articles/keyword/me")
@@ -31,18 +29,6 @@ interface ArticleAuthApi {
     suspend fun deleteKeyword(
         @Path("id") keywordId: Int
     ): Response<Unit>
-
-    /**
-     * 분실물 게시글 목록과 페이지 정보를 가져옴
-     * @param page 페이지 번호
-     * @param limit 페이지 당 게시글 수
-     */
-    @GET("articles/lost-item")
-    suspend fun fetchArticleLostAndFoundPagination(
-        @Query("page") page: Int,
-        @Query("limit") limit: Int,
-        @Query("type") type: String?
-    ): ArticleLostAndFoundPaginationResponse
 
     @POST("articles/lost-item")
     suspend fun uploadArticleLostAndFound(
